Code Review DF
Really nice job on this code! I learned a lot (especially from the regression-related plots) and thought there was a good deal of clean, elegant code to meet the objectives of the assignment. My only suggestion was including error bars on your geom_cols for visual analysis reasons (I provided a line of code for an example of this; I *think* I did it right). Great job!